requirements.txt
streamlit==0.73.0
pandas==1.0.5
numpy==1.18.5
newspaper3k==0.2.8